// src/types/bean.ts
export type Bean = {
    backgroundColor: string;
    beanId: number;
    colorGroup: string;
    description: string;
    flavorName: string;
    glutenFree: boolean;
    groupName: string[];
    imageUrl: string;
    ingredients: string[];
    kosher: boolean;
    seasonal: boolean;
    sugarFree: boolean;
};